fix(derive): hoist types out of traits
