+++ crates/derive/src/errors/mod.rs
@@ -7,9 +7,7 @@ mod stages;
 pub use stages::BatchDecompressionError;
 
 mod pipeline;
-pub use pipeline::{
-    PipelineEncodingError, PipelineError, PipelineErrorKind, PipelineResult, ResetError,
-};
+pub use pipeline::{PipelineEncodingError, PipelineError, PipelineErrorKind, ResetError};
 
 mod sources;
 pub use sources::{BlobDecodingError, BlobProviderError};
