+++ crates/derive/src/sources/ethereum.rs
@@ -2,9 +2,9 @@
 //! [DataAvailabilityProvider] trait for the Ethereum protocol.
 
 use crate::{
-    errors::PipelineResult,
     sources::{BlobSource, CalldataSource, EthereumDataSourceVariant},
     traits::{BlobProvider, ChainProvider, DataAvailabilityProvider},
+    types::PipelineResult,
 };
 use alloc::{boxed::Box, fmt::Debug};
 use alloy_primitives::{Address, Bytes};
