+++ crates/derive/src/test_utils/batch_stream.rs
@@ -3,9 +3,10 @@
 //! [`BatchStream`]: crate::stages::BatchStream
 
 use crate::{
-    errors::{PipelineError, PipelineResult},
+    errors::PipelineError,
     stages::BatchStreamProvider,
-    traits::{OriginAdvancer, OriginProvider, Signal, SignalReceiver},
+    traits::{OriginAdvancer, OriginProvider, SignalReceiver},
+    types::{PipelineResult, Signal},
 };
 use alloc::{boxed::Box, vec::Vec};
 use async_trait::async_trait;
