+++ crates/derive/src/traits/data_sources.rs
@@ -1,10 +1,7 @@
 //! Contains traits that describe the functionality of various data sources used in the derivation
 //! pipeline's stages.
 
-use crate::{
-    errors::{PipelineErrorKind, PipelineResult},
-    sources::IndexedBlobHash,
-};
+use crate::{errors::PipelineErrorKind, sources::IndexedBlobHash, types::PipelineResult};
 use alloc::{boxed::Box, fmt::Debug, string::ToString, vec::Vec};
 use alloy_eips::eip4844::Blob;
 use alloy_primitives::Bytes;
